Completed CSS for body in landing page exercise
The body section title, images, and text were formatted such that The
entire section was a flexbox going downwards which captured the title
and  content. The content itself was a flexbox going horizontally with
each of the items being contained within an image and subtext. Lastly,
the latter two were contained in their own flexbox going downward.
There was also image formatting completed.

 Changes to be committed:
	modified:   images/bus.jpeg
	modified:   images/face.jpeg
	modified:   images/old.jpeg
	modified:   images/park.jpeg
	modified:   index.html
	modified:   style.css